src/requirements.txt
absl-py==2.3.1
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
datasets==4.4.1
dill==0.4.0
editdistance==0.8.1
et_xmlfile==2.0.0
evaluate==0.4.6
exceptiongroup==1.3.1
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.8.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.2
idna==3.10
Janome==0.5.0
Jinja2==3.1.6
jiwer==4.0.0
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
Markdown==3.10
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
narwhals==1.39.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
omegaconf==2.3.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.2.1
propcache==0.4.1
protobuf==6.30.2
psutil==7.0.0
pyarrow==22.0.0
pydeck==0.9.1
pyngrok==7.2.8
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.14.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.24.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentencepiece==0.2.1
six==1.17.0
smmap==5.0.2
streamlit==1.45.1
sympy==1.13.1
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.21.1
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
Werkzeug==3.1.4
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
